Moved OPENRAVE_LOG4CXX into config.h. Fix a bunch of flags issues. Add logging.h.
