+++ CMakeLists.txt
@@ -46,7 +46,7 @@ option(OPT_BUILD_PACKAGE_DEFAULT "Set to ON to generate a default openrave packa
 option(OPT_IKFAST_FLOAT32 "Set to ON to allow loading of ikfast shared objects compiled with 32bit float (64bit double is always supported regardless of this option)" ON)
 option(OPT_FLANN "Temporary switch to force building of flann" OFF)
 option(OPT_CBINDINGS "Build the C-bindings libraries libopenrave_c and libopenrave-core_c" ON)
-option(OPT_LOG4CXX "Use log4cxx for logging" OFF)
+option(OPT_LOG4CXX "Use log4cxx for logging" ON)
 
 set(PACKAGE_VERSION "0" CACHE STRING "the package-specific version used for uploading the sources")
 set(CMAKE_MODULE_PATH "${CMAKE_CURRENT_SOURCE_DIR}/modules-cmake")
@@ -720,18 +720,29 @@ if( NOT FPARSER_FOUND )
 endif()
 
 # log4cxx logging lib
+set(OPENRAVE_LOG4CXX 0)
 set(OPENRAVE_LOG4CXX_INCLUDE_DIRS)
 set(OPENRAVE_LOG4CXX_LIB_DIRS)
 set(OPENRAVE_LOG4CXX_LIBRARY)
 if( OPT_LOG4CXX )
-  pkg_check_modules(LOG4CXX REQUIRED liblog4cxx)
+  pkg_check_modules(LOG4CXX liblog4cxx)
   if( LOG4CXX_FOUND )
-    set(OPENRAVE_EXPORT_CXXFLAGS "${OPENRAVE_EXPORT_CXXFLAGS} -DOPENRAVE_LOG4CXX")
-    set(OPENRAVE_LOG4CXX_INCLUDE_DIRS "-I${LOG4CXX_INCLUDEDIR}")
-    set(OPENRAVE_LOG4CXX_LIB_DIRS "-L${LOG4CXX_LIBDIR}")
-    set(OPENRAVE_LOG4CXX_LIBRARY "-llog4cxx")
+    if( MSVC )
+      set(OPENRAVE_LOG4CXX_INCLUDE_DIRS "/I\"${LOG4CXX_INCLUDEDIR}\"")
+      set(OPENRAVE_LOG4CXX_LIB_DIRS "/LIBPATH:\"${LOG4CXX_LIBDIR}\"")
+    else()
+      set(OPENRAVE_LOG4CXX_INCLUDE_DIRS "-I${LOG4CXX_INCLUDEDIR}")
+      set(OPENRAVE_LOG4CXX_LIB_DIRS "-L${LOG4CXX_LIBDIR}")
+    endif()
+    set(OPENRAVE_LOG4CXX_LIBRARY "-l${LOG4CXX_LIBRARIES}")
+
+    include_directories(${LOG4CXX_INCLUDEDIR})
+    set(OPENRAVE_LINK_DIRS ${OPENRAVE_LINK_DIRS} ${LOG4CXX_LIBDIR})
+    set(OPENRAVE_LOG4CXX 1)
   endif()
 endif()
+
+link_directories(${OPENRAVE_LINK_DIRS})
 
 message(STATUS "compiling local convexdecomposition library")
 add_subdirectory(3rdparty/convexdecomposition)
