+++ openrave-config.in
@@ -47,19 +47,19 @@ while test $# -gt 0; do
       echo -I@CMAKE_INSTALL_PREFIX@/include/@OPENRAVE_INCLUDE_INSTALL_DIR@ @OPENRAVE_BOOST_INCLUDE_DIRS@ @OPENRAVE_LOG4CXX_INCLUDE_DIRS@
       ;;
     --libs)
-      echo -L@CMAKE_INSTALL_PREFIX@/lib@LIB_SUFFIX@ -lopenrave@OPENRAVE_LIBRARY_SUFFIX@ @OPENRAVE_BOOST_LIB_DIRS@ @Boost_THREAD_LIBRARY_RELEASE@ @OPENRAVE_LOG4CXX_LIB_DIRS@ @OPENRAVE_LOG4CXX_LIBRARY@
+      echo -L@CMAKE_INSTALL_PREFIX@/lib@LIB_SUFFIX@ -lopenrave@OPENRAVE_LIBRARY_SUFFIX@ @OPENRAVE_BOOST_LIB_DIRS@ -l@Boost_THREAD_LIBRARY_RELEASE@ @OPENRAVE_LOG4CXX_LIB_DIRS@ @OPENRAVE_LOG4CXX_LIBRARY@
       ;;
     --libs-core)
-      echo -L@CMAKE_INSTALL_PREFIX@/lib@LIB_SUFFIX@ -lopenrave@OPENRAVE_LIBRARY_SUFFIX@ -lopenrave@OPENRAVE_LIBRARY_SUFFIX@-core @OPENRAVE_BOOST_LIB_DIRS@ @Boost_THREAD_LIBRARY_RELEASE@ @OPENRAVE_LOG4CXX_LIB_DIRS@ @OPENRAVE_LOG4CXX_LIBRARY@
+      echo -L@CMAKE_INSTALL_PREFIX@/lib@LIB_SUFFIX@ -lopenrave@OPENRAVE_LIBRARY_SUFFIX@ -lopenrave@OPENRAVE_LIBRARY_SUFFIX@-core @OPENRAVE_BOOST_LIB_DIRS@ -l@Boost_THREAD_LIBRARY_RELEASE@ @OPENRAVE_LOG4CXX_LIB_DIRS@ @OPENRAVE_LOG4CXX_LIBRARY@
       ;;
     --libs-only-l)
-      echo -lopenrave@OPENRAVE_LIBRARY_SUFFIX@ @Boost_THREAD_LIBRARY_RELEASE@ @OPENRAVE_LOG4CXX_LIBRARY@
+      echo -lopenrave@OPENRAVE_LIBRARY_SUFFIX@ -l@Boost_THREAD_LIBRARY_RELEASE@ @OPENRAVE_LOG4CXX_LIBRARY@
       ;;
     --libs-only-L)
       echo -L@CMAKE_INSTALL_PREFIX@/lib@LIB_SUFFIX@ @OPENRAVE_BOOST_LIB_DIRS@ @OPENRAVE_LOG4CXX_LIB_DIRS@
       ;;
     --shared-libs)
-      echo -L@CMAKE_INSTALL_PREFIX@/lib@LIB_SUFFIX@ -lopenrave@OPENRAVE_LIBRARY_SUFFIX@ @OPENRAVE_BOOST_LIB_DIRS@ @Boost_THREAD_LIBRARY_RELEASE@ @OPENRAVE_LOG4CXX_LIB_DIRS@ @OPENRAVE_LOG4CXX_LIBRARY@
+      echo -L@CMAKE_INSTALL_PREFIX@/lib@LIB_SUFFIX@ -lopenrave@OPENRAVE_LIBRARY_SUFFIX@ @OPENRAVE_BOOST_LIB_DIRS@ -l@Boost_THREAD_LIBRARY_RELEASE@ @OPENRAVE_LOG4CXX_LIB_DIRS@ @OPENRAVE_LOG4CXX_LIBRARY@
       ;;
     --python-dir)
       echo @CMAKE_INSTALL_PREFIX@/@OPENRAVE_PYTHON_INSTALL_DIR@
