+++ openrave-core.pc.in
@@ -10,5 +10,5 @@ Description: Open Robotics Automation Virtual Environment
 Version: @OPENRAVE_VERSION@
 URL: http://openrave.org
 Libs: -L${libdir} -lopenrave@OPENRAVE_LIBRARY_SUFFIX@ -lopenrave@OPENRAVE_LIBRARY_SUFFIX@-core
-Libs.private: @OPENRAVE_BOOST_LIB_DIRS@ @Boost_THREAD_LIBRARY@ @OPENRAVE_LOG4CXX_LIB_DIRS@ @OPENRAVE_LOG4CXX_LIBRARY@
+Libs.private: @OPENRAVE_BOOST_LIB_DIRS@ -l@Boost_THREAD_LIBRARY@ @OPENRAVE_LOG4CXX_LIB_DIRS@ @OPENRAVE_LOG4CXX_LIBRARY@
 Cflags: -I${includedir} @OPENRAVE_BOOST_INCLUDE_DIRS@ @OPENRAVE_LOG4CXX_INCLUDE_DIRS@ @OPENRAVE_EXPORT_CXXFLAGS@
