+++ plugins/configurationcache/CMakeLists.txt
@@ -10,15 +10,10 @@ install(TARGETS configurationcache DESTINATION ${OPENRAVE_PLUGINS_INSTALL_DIR} C
 if(OPT_PYTHON AND PYTHON_EXECUTABLE AND LAPACK_FOUND)
   if( Boost_PYTHON_FOUND AND Boost_THREAD_FOUND )
     if( HAVE_ALL_PYTHON_HEADERS )
-      set(OPENRAVEPY_CONFIGURATIONCACHE_COMPILE_FLAGS)
-      if( LOG4CXX_FOUND )
-        include_directories(${LOG4CXX_INCLUDEDIR})
-        set(OPENRAVEPY_CONFIGURATIONCACHE_COMPILE_FLAGS "${OPENRAVEPY_CONFIGURATIONCACHE_COMPILE_FLAGS} -DOPENRAVE_LOG4CXX")
-      endif()
       include_directories(${PYTHON_INCLUDE_PATH} ${PYTHON_INCLUDE_DIRS})
       add_library(openravepy_configurationcache SHARED openravepy_configurationcache.cpp)
       target_link_libraries(openravepy_configurationcache openravepy_int configurationcache ${PYTHON_LIBRARIES} ${Boost_PYTHON_LIBRARY} ${Boost_THREAD_LIBRARY} ${LAPACK_LIBRARIES})
-      set_target_properties(openravepy_configurationcache PROPERTIES PREFIX "" COMPILE_FLAGS "${OPENRAVEPY_CONFIGURATIONCACHE_COMPILE_FLAGS}")
+      set_target_properties(openravepy_configurationcache PROPERTIES PREFIX "" COMPILE_FLAGS "")
       if( WIN32 )
         set_target_properties(openravepy_configurationcache PROPERTIES SUFFIX ".pyd")
       endif()
