+++ python/bindings/CMakeLists.txt
@@ -42,14 +42,6 @@ if( Boost_PYTHON_FOUND AND Boost_THREAD_FOUND )
           set(OPENRAVEPY_COMPILE_FLAGS "${OPENRAVEPY_COMPILE_FLAGS} /bigobj")
         endif()
       endif()
-
-      if( LOG4CXX_FOUND )
-        include_directories(${LOG4CXX_INCLUDEDIR})
-        set(OPENRAVEPY_COMPILE_FLAGS "${OPENRAVEPY_COMPILE_FLAGS} -DOPENRAVE_LOG4CXX")
-        set(ANN_CFLAGS "${ANN_CFLAGS} -DOPENRAVE_LOG4CXX")
-        set(CONVEXDECOMPOSITION_CFLAGS "${CONVEXDECOMPOSITION_CFLAGS} -DOPENRAVE_LOG4CXX")
-      endif()
-
       ## openrave bindings
       add_library(openravepy_int SHARED openravepy_kinbody.cpp openravepy_int.cpp openravepy_robot.cpp openravepy_int.h openravepy_global.cpp openravepy_collisionchecker.cpp openravepy_controller.cpp openravepy_ikparameterization.cpp openravepy_iksolver.cpp openravepy_module.cpp openravepy_physicsengine.cpp openravepy_planner.cpp openravepy_planningutils.cpp openravepy_sensor.cpp openravepy_sensorsystem.cpp openravepy_spacesampler.cpp openravepy_trajectory.cpp openravepy_viewer.cpp bindings.cpp bindings.h docstrings.cpp docstrings.h)
       # stdc++ has to be included before opengl libraries due to some ATI bug (http://wiki.fifengine.de/Segfault_in_cxa_allocate_exception#Workaround)
